+++ packages/cli/test/projects/fixtures.ts
@@ -61,8 +61,7 @@ export const myProject_v1: Provisioner.Project = {
 
 export const myProject_yaml = `id: my-project
 name: My Project
-cli:
-  version: 2
+schema_version: '4.0'
 description: my lovely project
 collections: []
 credentials: []
@@ -81,7 +80,7 @@ workflows:
       - id: transform-data
         name: Transform data
         expression: fn()
-        adaptor: "@openfn/language-common@latest"
+        adaptor: '@openfn/language-common@latest'
         openfn:
           uuid: 66add020-e6eb-4eec-836b-20008afca816
       - id: webhook